add working osm_liberty_style_edit and config
the two jsons now work correctly with the docker run command, by
specifying the correct sources url using "mbtiles://" and also the
sprite and glyphs paths in the style json.

for the folder structure, the style and config json should be with the
mbtiles, and the sprite folder provided by osm-liberty should be placed
in that folder i.e.

- tiles
- config.json
- style.json
- sprites/

confirmed that this is working and serving tiles with the new style.

obtained all the required osm-liberty things from the github. refer to
the original version to see the edits (namely removal of the required
mapbox keys)